Adds functions to test for link correctness, validity, and visibility
When you add new tests, there's a lot of reused code that I thought
could be turned into functions. This should reduce some of the code.

I also renamed the tests for the footers to match the naming scheme of
the other tests on for the page.

I'm adding it to only this test to see if this method is ok to use
before adding it to other pages of the test.